fix: InputView OutputView 분리.
